pyproject.toml
[tool.poetry]
name = "agent-transpiler"
version = "0.1.0"
description = ""
authors = ["diegoasua <baxuaa@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
anthropic = "^0.25.7"
python-dotenv = "^1.0.1"
flake8 = "^7.0.0"
black = "^24.4.2"
autopep8 = "^2.1.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
